requirements.txt
pip~=22.3.1
wheel~=0.38.4
certifi~=2022.12.7
requests~=2.28.1
idna~=3.4
urllib3~=1.26.13
setuptools~=65.6.3
python-dotenv~=0.21.0
pyTelegramBotAPI~=4.8.0